# app/public/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CityCatalyst API",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/v0/inventory/{inventory}/hiap": {
      "get": {
        "tags": [
          "Inventory HIAP"
        ],
        "summary": "Get HIAP ranking or related data for an inventory.",
        "description": "Returns HIAP insights for the selected actionType and language. Requires a signed‑in user with access to the inventory. Response is wrapped in { data } (actionType‑dependent shape).",
        "parameters": [
          {
            "in": "path",
            "name": "inventory",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "actionType",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lng",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HIAP result wrapped in data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Inventory HIAP"
        ],
        "summary": "Update selection status of ranked actions.",
        "description": "Updates the isSelected field for ranked actions. All actions not in the selectedActionIds array will be set to false.",
        "parameters": [
          {
            "in": "path",
            "name": "inventory",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "selectedActionIds"
                ],
                "properties": {
                  "selectedActionIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Selection updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "updated": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}